config.py
import os

from dotenv import load_dotenv

load_dotenv()


class Config:
    SECRET_KEY = os.getenv("SECRET_KEY")
    STATIC_FOLDER = "static"
    TEMPLATES_FOLDER = "templates"

    # JWT
    ENCRYPTION_ALGO = "HS256"
    ACCESS_TOKEN_EXPIRE_MINUTES = 30

    # Assets
    ASSETS_DEBUG = True

    # Database
    SQLALCHEMY_TRACK_MODIFICATIONS = False
    SQLALCHEMY_ECHO = True


class Development(Config):
    FLASK_ENV = "development"
    ENV = "development"
    DEBUG = True
    TESTING = True
    SQLALCHEMY_DATABASE_URI = "mysql+pymysql://root:password@db/app"
