requirements.txt
flask==1.1.2
flask-sqlalchemy==2.4.4
flask-marshmallow==0.14.0
flask-assets==2.0
marshmallow-sqlalchemy==0.24.1
isort==5.6.4
black
cssmin==0.2.0
jsmin==2.2.2
argon2-cffi==19.2.0
pyjwt==1.7.1
pymysql==0.9.3
python-dotenv==0.14.0